Backend-Orasync/requirements.txt
<<<<<<< HEAD
Flask==3.1.0
Flask-Cors==5.0.1
firebase-admin==6.4.0
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==21.2.0
google-genai>=1.13.0
google-cloud-texttospeech==2.18.0
google-cloud-storage==2.14.0
=======
Flask>=2.3.0
Flask-Cors>=4.0.0
google-generativeai>=0.5.0
google-cloud-storage>=2.10.0
google-cloud-texttospeech>=2.16.0
Pillow>=10.2.0
python-dotenv>=1.0.0
firebase-admin>=6.5.0
gunicorn>=21.0.0
psutil>=5.9.0
requests>=2.31.0
opencv-python>=4.8.0
numpy>=1.26.0
google-genai==1.24.0
>>>>>>> 9129cfe4b41d693ce0501e8a686c17ac643b01c0
